Make the `results` goal always scan jtl files in the "results" directory
The field `resultFilesLocations` is never set except in `RunJMeterMojo`. However
this can't be used by `CheckResultsMojo` since it's a completely different
execution. Make the results goal always scan all jtl files.

One potential drawback is that it will scan result files that were not generated
by the same maven execution. That said it can also be considered an advantage if
the goal needs to be run later. If not then the recommendation is run clean
before running the tests.

Fixes #190